[KOGITO-9811] Moving expression from ForEach to Process
